Make CUPS run as non-root user
In this case, we do not want CUPS to run as UID 0, as this is:

A) unsafe
B) fails by default in OpenShift environments

Move everything to UID 1001 and similarly, listen on port 6631,
and update the services/endpoints accordingly.

Massive thanks to Mason Loring Bliss for help in making CUPS
cooperate nicely.